trash-identification/requirements.txt
torch
torchvision
opencv-python
numpy
pillow
django
djangorestframework
